Product Decisions lens with workspace-wide marker
- DecisionsIndex extracted to a shared component (filters + table); the
  workspace page becomes a thin wrapper around it
- products/[productSlug]/decisions: the same index locked to the product's
  repos PLUS workspace-level (null-product) ADRs — adr.list gains
  includeWorkspaceWide for the union; null-product rows render a
  'workspace-wide' badge in every lens per the PRD's open-question default
- product layout gains a Decisions tab (IconGavel) between Insights and
  Retro
